Refactor: future_results, io.async. -> ppr.
